ПР14_1.0

+++ controllers/cards.js
@@ -1,45 +1,53 @@
 const Card = require('../models/card');
 const { BadRequest, NotFound, ServerError } = require('../utils/constants');
+const NotFoundError = require('../errors/NotFoundError'); // 404
+const BadRequestError = require('../errors/BadRequestError'); // 400
+const ConflictError = require('../errors/ConflictError'); // 409
+const UnauthorizedError = require('../errors/UnauthorizedError'); // 401
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => {
       res.send({ data: cards });
     })
-    .catch((err) => {
-      res.status(ServerError).send({ message: err.message });
-    });
+    .catch(next);
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   const userId = req.user._id;
   Card.create({ name, link, owner: userId })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(BadRequest).send({ message: err.message });
-        return;
+        next(new BadRequestError(err.message));
       }
-      res.status(ServerError).send({ message: 'Произошла ошибка' });
     });
 };
 
 const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (card === null) {
         res.status(NotFound).send({ message: 'Карточка с указанным _id не найдена' });
         return;
       }
-      res.send({ data: card });
+      if (card.owner === req.user._id) {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((cardDelete) => {
+            res.send({ data: cardDelete });
+          })
+          .catch((err) => {
+            if (err.name === 'CastError') {
+              res.status(BadRequest).send({ message: err.message });
+              return;
+            }
+            res.status(ServerError).send({ message: err.message });
+          });
+      }
     })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(BadRequest).send({ message: err.message });
-        return;
-      }
-      res.status(ServerError).send({ message: err.message });
+      res.status(NotFound).send({ message: err.message })
     });
 };
